+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 * nostr: add `CowTag` ([Yuki Kishimoto])
 * nostr: add `EventBorrow` ([Yuki Kishimoto])
 * database: add `Events::first_owned` and `Events::last_owned` ([Yuki Kishimoto])
+* database: impl `FlatBufferDecodeBorrowed` for `EventBorrow` ([Yuki Kishimoto])
 * pool: add `Relay::try_connect` ([Yuki Kishimoto])
 * pool: add `Relay::wait_for_connection` ([Yuki Kishimoto])
 * pool: add `RelayPool::try_connect` ([Yuki Kishimoto])